Move _hamming_neighbors function to BooleanSpace method
